pyproject.toml
[project]
name = "yugioh-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "darellchua",email = "darellchua2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlsxwriter (>=3.2.0,<4.0.0)",
    "boto3 (>=1.36.2,<2.0.0)"
]

[tool.poetry]
package-mode = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

